refactor: unexport CalculatePaymentAmount, add zero-value tests
- Rename CalculatePaymentAmount to calculatePaymentCoin (unexported)
  since it's only used within the keeper package.
- Move test to package keeper (internal) to access the unexported func.
- Add zero_blob_size and zero_gas_per_blob_byte test cases.
