+++ x/fibre/keeper/msg_server.go
@@ -298,12 +298,12 @@ func (ms msgServer) UpdateFibreParams(goCtx context.Context, msg *types.MsgUpdat
 func (ms msgServer) calculatePaymentAmount(ctx sdk.Context, blobSize uint32) sdk.Coin {
 	params := ms.GetParams(ctx)
 	// TODO: this assumes 1 utia per gas which may not be correct.
-	return CalculatePaymentAmount(blobSize, params.GasPerBlobByte)
+	return calculatePaymentCoin(blobSize, params.GasPerBlobByte)
 }
 
-// CalculatePaymentAmount computes the payment coin from blobSize and gasPerBlobByte.
+// calculatePaymentCoin computes the payment coin from blobSize and gasPerBlobByte.
 // Both operands are widened to uint64 before multiplication to prevent uint32 overflow.
-func CalculatePaymentAmount(blobSize, gasPerBlobByte uint32) sdk.Coin {
+func calculatePaymentCoin(blobSize, gasPerBlobByte uint32) sdk.Coin {
 	result := uint64(blobSize) * uint64(gasPerBlobByte)
 	return sdk.NewCoin(appconsts.BondDenom, math.NewIntFromUint64(result))
 }
